pyproject.toml
[project]
name = "aquaforge"
version = "1.0.0-next"
description = "Swim meet optimization platform for competitive swimming teams"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "reflex>=0.8.24",
    "reflex-hosting-cli>=0.1.61",
    "werkzeug>=3.0.0",
    # Data processing
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "openpyxl>=3.1.2",
    # Optimization
    "gurobipy>=11.0.0",
    # PDF
    "pypdf>=4.2.0",
    "pdfplumber>=0.11.0",
    "pdfminer.six>=20231228",
    "reportlab",
    "xhtml2pdf>=0.2.16",
    "svglib>=1.6.0",
    "html5lib>=1.1",
    "lxml>=5.2.0",
    # Backend
    "granian>=2.6.0",
    "starlette>=0.37.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-socketio>=5.11.0",
    "python-engineio>=4.9.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "urllib3>=2.2.0",
    "beautifulsoup4>=4.12.0",
    "watchfiles>=0.21.0",
    "websockets>=12.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.30",
    "sqlmodel>=0.0.19",
    "alembic>=1.13.0",
    # Utils
    "python-dateutil>=2.9.0",
    "pyyaml>=6.0.1",
    "pillow>=10.3.0",
    "psutil>=5.9.0",
    "packaging>=24.0",
    "rich>=13.7.0",
    "typing_extensions>=4.12.0",
    "wrapt>=1.16.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.3.0",
    "anyio>=4.0.0",
    # Code quality
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    # Development tools
    "ipython>=8.0.0",
    "watchdog>=4.0.0",
    # Documentation
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    # Debugging
    "debugpy>=1.8.0",
    # Scraping / E2E
    "playwright>=1.40.0",
]
prod = [
    # Configuration
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    # Logging & observability
    "structlog>=24.0.0",
    "python-json-logger>=2.0.0",
    "sentry-sdk>=1.40.0",
    # Async & background jobs
    "arq>=0.26.0",
    "aioredis>=2.0.0",
    # HTTP client
    "tenacity>=8.2.0",
    "aiohttp>=3.9.0",
    # Security & auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.0",
    # Database (async)
    "asyncpg>=0.29.0",
    "databases>=0.9.0",
    # Caching
    "aiocache>=0.12.0",
    # Metrics & tracing
    "prometheus-client>=0.20.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
    "opentelemetry-instrumentation-redis>=0.45b0",
    "opentelemetry-exporter-otlp>=1.24.0",
    # Rate limiting
    "slowapi>=0.1.9",
    "limits>=3.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["swim_ai_reflex*"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "N806", "N803", "N805", "UP008", "UP038"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: deselect with '-m \"not slow\"'",
    "timeout",
    "e2e: requires running servers",
    "integration",
]
asyncio_mode = "auto"
addopts = "--tb=short -q --strict-markers"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
